refs #17 : correct long cast syntax. no ticket : add drop schema to clear script, to harmonize with create script where create schema has been previously added

+++ chouette/src/main/java/fr/certu/chouette/service/importateur/multilignes/hastus/impl/LecteurCourse.java
@@ -106,7 +106,7 @@ public class LecteurCourse implements ILecteurCourse {
 		String key = sdf1.format(debut)+bigInt.toString()+sdf1.format(fin);
 		TableauMarche tableauMarche = tableauMarcheParValeur.get(key);
 		if (tableauMarche == null) {
-			long numberOfDays = (long)1 + (fin.getTime()/((long)24*(long)60*(long)60*(long)1000) - debut.getTime()/((long)24*(long)60*(long)60*(long)1000));
+			long numberOfDays = (long)1 + (fin.getTime()/(24l*60l*60l*1000l) - debut.getTime()/(24l*60l*60l*1000l));
 			if (joursValides.length() != (int)numberOfDays)
 				throw new ServiceException(CodeIncident.INVALIDE_DATESJOURSVALIDITE_COURSE, "Le champs jours valides ("+joursValides+") n'est pas conforme aux dates de debut ("+ligneCSV[11].trim()+") est de fin ("+ligneCSV[12].trim()+") : "+(int)numberOfDays+" jours.");
 			long time = debut.getTime();
@@ -114,10 +114,13 @@ public class LecteurCourse implements ILecteurCourse {
 			Date last = null;
 			String value = null;
 			tableauMarche = new TableauMarche();
-			for (int i = 0; i < numberOfDays; i++) {
+			for (int i = 0; i < numberOfDays; i++) 
+			{
 				Date date = new Date(time);
-				if (joursValides.charAt(i) == '1') {
-					if (first == null) {
+				if (joursValides.charAt(i) == '1') 
+				{
+					if (first == null) 
+					{
 						first = new Date(time);
 						value = "1";
 					}
@@ -127,6 +130,7 @@ public class LecteurCourse implements ILecteurCourse {
 				if (value != null)
 					value += joursValides.charAt(i);
 				time += (long)24*60*60*1000;
+				
 			}
 			if (value == null)
 				return; // TM vide
@@ -134,7 +138,8 @@ public class LecteurCourse implements ILecteurCourse {
 			tableauMarche.setCreationTime(new Date(System.currentTimeMillis()));
 			tableauMarcheParValeur.put(key, tableauMarche);
 			int index = value.length()-1;
-			while (value.charAt(index) == '0') {
+			while (value.charAt(index) == '0') 
+			{
 				value = value.substring(0, index);
 				index = value.length()-1;
 			}
